requirements.txt
Django==4.1
psycopg2==2.9.5
djangorestframework==3.14.0
pillow==9.4.0
django-filter==22.1
pytest==7.2.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
drf-spectacular==0.25.1
